feat: split image upload component
