+++ src/components/ImageUpload/index.tsx
@@ -4,27 +4,18 @@ import {
   ModalOverlay,
   ModalContent,
   Input,
-  Button,
   ModalBody,
   ModalCloseButton,
-  ModalFooter,
   ModalHeader,
   Box,
   Center,
-  Flex,
-  Slider,
-  SliderFilledTrack,
-  SliderThumb,
-  SliderTrack,
-  Text,
   useToast,
   Spinner,
 } from "@chakra-ui/react";
-import { useCallback, useState } from "react";
+import { useState } from "react";
 import { useDispatch } from "react-redux";
-import Cropper, { Area } from "react-easy-crop";
-import { getCroppedImg } from "@/utils/manageExternalImage";
 import { AnyAction } from "@reduxjs/toolkit";
+import CropImage from "./CropImage";
 
 interface IImageUpload {
   name: string;
@@ -40,21 +31,10 @@ const ImageUpload = (props: IImageUpload) => {
   const dispatch = useDispatch();
   const toast = useToast();
   const [picture, setPicture] = useState("");
-  const [crop, setCrop] = useState({ x: 0, y: 0 });
-  const [zoom, setZoom] = useState(1);
-  const [rotation, setRotation] = useState(0);
-  const [croppedAreaPixels, setCroppedAreaPixels] = useState<Area | null>(null);
-  const [mediaSize, setMediaSize] = useState({
-    width: 0,
-    height: 0,
-    naturalWidth: 0,
-    naturalHeight: 0,
-  });
   const [loading, setLoading] = useState(false);
 
   const onClose = () => {
     setPicture("");
-    handleReset();
     dispatch(switchUpload(false));
   };
 
@@ -70,50 +50,10 @@ const ImageUpload = (props: IImageUpload) => {
       setLoading(false);
       return;
     }
-    handleReset();
     const file = e.target.files[0];
     imgSet(file);
     setLoading(false);
   };
-
-  const handleReset = () => {
-    setCrop({ x: 0, y: 0 });
-    setZoom(1);
-    setRotation(0);
-
-    const croppedAreaPixelsValue = {
-      width: 0,
-      height: 0,
-      x: 0,
-      y: 0,
-    };
-    if (mediaSize.naturalWidth >= mediaSize.naturalHeight) {
-      croppedAreaPixelsValue.width = mediaSize.naturalHeight;
-      croppedAreaPixelsValue.height = mediaSize.naturalHeight;
-      croppedAreaPixelsValue.x = (mediaSize.naturalWidth - mediaSize.naturalHeight) / 2;
-    } else {
-      croppedAreaPixelsValue.width = mediaSize.naturalWidth;
-      croppedAreaPixelsValue.height = mediaSize.naturalWidth;
-      croppedAreaPixelsValue.y = (mediaSize.naturalHeight - mediaSize.naturalWidth) / 2;
-    }
-
-    setCroppedAreaPixels(croppedAreaPixelsValue);
-  };
-
-  const onCropComplete = useCallback((croppedArea: Area, croppedAreaPixels: Area) => {
-    setCroppedAreaPixels(croppedAreaPixels);
-  }, []);
-
-  const handleApply = useCallback(async () => {
-    if (!picture || !croppedAreaPixels) return;
-    const cropCanvas = (await getCroppedImg(
-      picture,
-      croppedAreaPixels,
-      rotation
-    )) as HTMLCanvasElement;
-    setImage(cropCanvas.toDataURL(), cropCanvas.width, cropCanvas.height);
-    onClose();
-  }, [picture, croppedAreaPixels, rotation]);
 
   return (
     <Modal isOpen={isOpen} onClose={onClose} size="lg" isCentered returnFocusOnClose={false}>
@@ -138,6 +78,7 @@ const ImageUpload = (props: IImageUpload) => {
               onChange={handleInputImage}
               width="100%"
               height="100%"
+              aria-label="upload file"
             ></Input>
             <Box position="absolute" color="white" fontWeight="500" pointerEvents="none">
               Choose an image
@@ -149,89 +90,9 @@ const ImageUpload = (props: IImageUpload) => {
             </Center>
           )}
           {picture && !loading && (
-            <>
-              <Center position="relative" height="260px" marginTop="10px">
-                <Cropper
-                  image={picture}
-                  crop={crop}
-                  rotation={rotation}
-                  zoom={zoom}
-                  cropShape={"round"}
-                  aspect={1 / 1}
-                  onCropChange={setCrop}
-                  onRotationChange={setRotation}
-                  onCropComplete={onCropComplete}
-                  onZoomChange={setZoom}
-                  setMediaSize={setMediaSize}
-                />
-              </Center>
-              <Flex
-                justifyContent="space-between"
-                alignContent="center"
-                margin="0 0.5rem"
-                marginTop="15px"
-              >
-                <Box minWidth="120px">
-                  <Text fontSize="sm" color="brand.primary" fontWeight="600">
-                    Zoom
-                  </Text>
-                  <Slider
-                    aria-label="zoom-slider"
-                    value={zoom}
-                    min={1}
-                    max={3}
-                    step={0.1}
-                    aria-labelledby="Zoom"
-                    onChange={(zoom) => setZoom(Number(zoom))}
-                  >
-                    <SliderTrack>
-                      <SliderFilledTrack />
-                    </SliderTrack>
-                    <SliderThumb />
-                  </Slider>
-                </Box>
-                <Box minWidth="120px">
-                  <Text fontSize="sm" color="brand.primary" fontWeight="600">
-                    Rotation
-                  </Text>
-                  <Slider
-                    aria-label="rotate-slider"
-                    value={rotation}
-                    min={-180}
-                    max={180}
-                    step={0.5}
-                    aria-labelledby="rotate"
-                    onChange={(rotation) => setRotation(Number(rotation))}
-                  >
-                    <SliderTrack>
-                      <SliderFilledTrack />
-                    </SliderTrack>
-                    <SliderThumb />
-                  </Slider>
-                </Box>
-                <Box>
-                  <Button
-                    colorScheme="blue"
-                    variant="outline"
-                    height="33px"
-                    width="60px"
-                    marginTop="2px"
-                    onClick={handleReset}
-                  >
-                    Reset
-                  </Button>
-                </Box>
-              </Flex>
-            </>
+            <CropImage name={name} picture={picture} setImage={setImage} onClose={onClose} />
           )}
         </ModalBody>
-        <ModalFooter justifyContent="center">
-          {picture && !loading && (
-            <Button variant="shareBtn" onClick={handleApply}>
-              Apply {name}
-            </Button>
-          )}
-        </ModalFooter>
       </ModalContent>
     </Modal>
   );
